requirements.txt
Flask
python-dotenv
agno
firecrawl-sdk
# Añade aquí otras dependencias directas que tu proyecto necesite 